Stabilized the Configuration system
Stabilized the Configuration system by adding safeguards in case the
robot is started with out the appropriate config settings set. The safe
gaurds prevent errors from ending the robot programs, so that the
configuration file can be rewritten with the appropriate settings can be
added.
ConfigNotFoundException was added to handle when the a settings does not
exist.
The Robot class also has new init methods to avoid confusion, and making
tracking down errors easier.